# Manual control of MAC cluster for improved interwave performance