To avoid register spills seen on latest hipclang, getting best perf requires patches yet to be upstreamed + internal LLVM based off Mainline #9110 as listed here:
http://gerrit-git.amd.com/plugins/gitiles/lightning/ec/llvm-project/+/refs/heads/aukerbow/9110-ck

### Caveat
There is known verification in some `DeviceGemmXdl` kernels that have register spills. `DeviceGemmXdlCShuffle` kernels passed verifications.

List of kernels failed verification due to suspected register spill, using `./ckProfiler gemm`. A lot of false positives from BF16 verification failure is due to rounding errors

Update (5/5/2022): The latest commit 56790a62c avoids validation failure by rolling out interwave scheduling to C-shuffle gemm variants only.

<details><summary> Detail </summary>
<p>

test fp32 TT
test fp32 TN
test fp32 NT
test fp32 NN
test fp16 TT
Perf:   0.693763 ms, 1.54771 TFlops, 12.2804 GB/s, DeviceGemmXdl<256, 256, 128, 4, 8, 32, 32, 4, 2>
   max err: 4049
Perf:   0.633443 ms, 1.695088 TFlops, 13.4498 GB/s, DeviceGemmXdl<256, 128, 256, 4, 8, 32, 32, 2, 4>
   max err: 4311
Perf:   0.541443 ms, 1.983112 TFlops, 15.73514 GB/s, DeviceGemmXdl<128, 128, 128, 4, 8, 32, 32, 4, 2>
   max err: 4070
test fp16 TN
Perf:   0.425282 ms, 2.52478 TFlops, 20.033 GB/s, DeviceGemmXdl<128, 128, 128, 4, 8, 32, 32, 4, 2>
   max err: 4070
test fp16 NT
Perf:   0.577923 ms, 1.85793 TFlops, 14.7419 GB/s, DeviceGemmXdl<256, 256, 128, 4, 8, 32, 32, 4, 2>
   max err: 4049
Perf:   0.485923 ms, 2.209695 TFlops, 17.53298 GB/s, DeviceGemmXdl<128, 128, 128, 4, 8, 32, 32, 4, 2>
   max err: 4070
test fp16 NN
Perf:   0.546723 ms, 1.96396 TFlops, 15.5832 GB/s, DeviceGemmXdl<256, 256, 128, 4, 8, 32, 32, 4, 2>
   max err: 4049
Perf:   0.468322 ms, 2.292742 TFlops, 18.19193 GB/s, DeviceGemmXdl<128, 128, 128, 4, 8, 32, 32, 4, 2>
   max err: 4070
test bf16 TT
Perf:   0.723845 ms, 1.48339 TFlops, 11.77 GB/s, DeviceGemm_Xdl_CShuffle<256, 256, 128, 32, 8, 2>
   max err: 16
Perf:   0.763046 ms, 1.407178 TFlops, 11.16536 GB/s, DeviceGemm_Xdl_CShuffle<256, 256, 128, 32, 8, 8>
   max err: 16
Perf:   0.676324 ms, 1.587615 TFlops, 12.59704 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 256, 32, 8, 2>
   max err: 16
Perf:   0.674564 ms, 1.591757 TFlops, 12.62991 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 256, 32, 8, 8>
   max err: 16
Perf:   0.548483 ms, 1.957657 TFlops, 15.53317 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 128, 32, 8, 2>
   max err: 16
Perf:   0.564003 ms, 1.903787 TFlops, 15.10574 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 128, 32, 8, 8>
   max err: 16
Perf:   0.374082 ms, 2.870338 TFlops, 22.7749 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 128, 32, 8, 2>
   max err: 16
Perf:   0.390722 ms, 2.748097 TFlops, 21.80497 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 128, 32, 8, 8>
   max err: 16
Perf:   0.287841 ms, 3.73033 TFlops, 29.59856 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 64, 32, 8, 2>
   max err: 16
Perf:   0.336962 ms, 3.186537 TFlops, 25.2838 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 64, 32, 8, 8>
   max err: 16
Perf:   0.293601 ms, 3.657146 TFlops, 29.01788 GB/s, DeviceGemm_Xdl_CShuffle<128, 64, 128, 32, 8, 2>
   max err: 16
Perf:   0.307042 ms, 3.497052 TFlops, 27.7476 GB/s, DeviceGemm_Xdl_CShuffle<128, 64, 128, 32, 8, 8>
   max err: 16
Perf:   0.229441 ms, 4.679817 TFlops, 37.13234 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 64, 32, 8, 2>
   max err: 16
Perf:   0.282402 ms, 3.802175 TFlops, 30.16862 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 64, 32, 8, 8>
   max err: 16
Perf:   0.224641 ms, 4.779812 TFlops, 37.92576 GB/s, DeviceGemm_Xdl_CShuffle<256, 64, 128, 32, 8, 2>
   max err: 16
Perf:   0.236802 ms, 4.534345 TFlops, 35.97807 GB/s, DeviceGemm_Xdl_CShuffle<256, 64, 128, 32, 8, 8>
   max err: 16
test bf16 TN
Perf:   0.504483 ms, 2.1284 TFlops, 16.8879 GB/s, DeviceGemm_Xdl_CShuffle<256, 256, 128, 32, 8, 8>
   max err: 16
Perf:   0.565125 ms, 1.900008 TFlops, 15.07574 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 256, 32, 8, 8>
   max err: 16
Perf:   0.443842 ms, 2.419198 TFlops, 19.1953 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 128, 32, 8, 8>
   max err: 16
Perf:   0.282242 ms, 3.80433 TFlops, 30.18573 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 128, 32, 8, 8>
   max err: 16
Perf:   0.244321 ms, 4.3948 TFlops, 34.87085 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 64, 32, 8, 8>
   max err: 16
Perf:   0.260962 ms, 4.114552 TFlops, 32.64721 GB/s, DeviceGemm_Xdl_CShuffle<128, 64, 128, 32, 8, 8>
   max err: 16
Perf:   0.249121 ms, 4.310122 TFlops, 34.19896 GB/s, DeviceGemm_Xdl_CShuffle<64, 64, 64, 32, 8, 8>
   max err: 16
Perf:   0.178082 ms, 6.02948 TFlops, 47.84133 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 64, 32, 8, 8>
   max err: 16
Perf:   0.211681 ms, 5.072453 TFlops, 40.24773 GB/s, DeviceGemm_Xdl_CShuffle<256, 64, 128, 32, 8, 8>
   max err: 16
Perf:   0.167681 ms, 6.40348 TFlops, 50.80886 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 32, 32, 8, 8>
   max err: 16
Perf:   0.218881 ms, 4.905596 TFlops, 38.9238 GB/s, DeviceGemm_Xdl_CShuffle<128, 32, 128, 32, 8, 8>
   max err: 16
Perf:   0.158561 ms, 6.77179 TFlops, 53.73124 GB/s, DeviceGemm_Xdl_CShuffle<64, 64, 32, 32, 8, 8>
   max err: 16
Perf:    0.15584 ms, 6.890028 TFlops, 54.66941 GB/s, DeviceGemm_Xdl_CShuffle<64, 32, 64, 32, 8, 8>
   max err: 16
test bf16 NT
Perf:   0.579044 ms, 1.85434 TFlops, 14.7134 GB/s, DeviceGemm_Xdl_CShuffle<256, 256, 128, 32, 2, 2>
   max err: 16
Perf:   0.629445 ms, 1.705855 TFlops, 13.53522 GB/s, DeviceGemm_Xdl_CShuffle<256, 256, 128, 32, 8, 8>
   max err: 16
Perf:   0.579203 ms, 1.853826 TFlops, 14.70932 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 256, 32, 2, 2>
   max err: 16
Perf:   0.600004 ms, 1.789558 TFlops, 14.19937 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 256, 32, 8, 8>
   max err: 16
Perf:   0.540803 ms, 1.985458 TFlops, 15.75376 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 128, 32, 2, 2>
   max err: 16
Perf:   0.538403 ms, 1.994309 TFlops, 15.82398 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 128, 32, 8, 8>
   max err: 16
Perf:   0.311842 ms, 3.443224 TFlops, 27.3205 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 128, 32, 2, 2>
   max err: 16
Perf:   0.325282 ms, 3.300957 TFlops, 26.19167 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 128, 32, 8, 8>
   max err: 16
Perf:   0.348642 ms, 3.079783 TFlops, 24.43676 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 64, 32, 2, 2>
   max err: 16
Perf:   0.326402 ms, 3.28963 TFlops, 26.1018 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 64, 32, 8, 8>
   max err: 16
Perf:   0.306082 ms, 3.50802 TFlops, 27.83463 GB/s, DeviceGemm_Xdl_CShuffle<128, 64, 128, 32, 2, 2>
   max err: 16
Perf:   0.320002 ms, 3.355422 TFlops, 26.62383 GB/s, DeviceGemm_Xdl_CShuffle<128, 64, 128, 32, 8, 8>
   max err: 16
Perf:   0.198081 ms, 5.420721 TFlops, 43.01109 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 64, 32, 2, 2>
   max err: 16
Perf:   0.216161 ms, 4.967325 TFlops, 39.41359 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 64, 32, 8, 8>
   max err: 16
Perf:   0.195521 ms, 5.491696 TFlops, 43.57425 GB/s, DeviceGemm_Xdl_CShuffle<256, 64, 128, 32, 2, 2>
   max err: 16
Perf:   0.208961 ms, 5.13848 TFlops, 40.77163 GB/s, DeviceGemm_Xdl_CShuffle<256, 64, 128, 32, 8, 8>
   max err: 16
test bf16 NN
Perf:   0.533923 ms, 2.01104 TFlops, 15.9568 GB/s, DeviceGemm_Xdl_CShuffle<256, 256, 128, 32, 2, 8>
   max err: 16
Perf:   0.592325 ms, 1.812758 TFlops, 14.38346 GB/s, DeviceGemm_Xdl_CShuffle<256, 256, 128, 32, 8, 8>
   max err: 16
Perf:   0.530723 ms, 2.023168 TFlops, 16.05297 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 256, 32, 2, 8>
   max err: 16
Perf:   0.523203 ms, 2.052247 TFlops, 16.2837 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 256, 32, 8, 8>
   max err: 16
Perf:   0.480323 ms, 2.235458 TFlops, 17.7374 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 128, 32, 2, 8>
   max err: 16
Perf:   0.509443 ms, 2.107678 TFlops, 16.72352 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 128, 32, 8, 8>
   max err: 16
Perf:   0.305921 ms, 3.509866 TFlops, 27.84928 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 128, 32, 2, 8>
   max err: 16
Perf:   0.307362 ms, 3.493411 TFlops, 27.71872 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 128, 32, 8, 8>
   max err: 16
Perf:   0.278881 ms, 3.850179 TFlops, 30.54952 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 64, 32, 2, 8>
   max err: 16
Perf:   0.301922 ms, 3.556355 TFlops, 28.21815 GB/s, DeviceGemm_Xdl_CShuffle<128, 128, 64, 32, 8, 8>
   max err: 16
Perf:   0.265921 ms, 4.037823 TFlops, 32.03839 GB/s, DeviceGemm_Xdl_CShuffle<128, 64, 128, 32, 2, 8>
   max err: 16
Perf:   0.276962 ms, 3.876856 TFlops, 30.76119 GB/s, DeviceGemm_Xdl_CShuffle<128, 64, 128, 32, 8, 8>
   max err: 16
Perf:   0.189761 ms, 5.658391 TFlops, 44.8969 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 64, 32, 2, 8>
   max err: 16
Perf:   0.198241 ms, 5.416346 TFlops, 42.97638 GB/s, DeviceGemm_Xdl_CShuffle<256, 128, 64, 32, 8, 8>
   max err: 16
Perf:   0.197441 ms, 5.438292 TFlops, 43.15051 GB/s, DeviceGemm_Xdl_CShuffle<256, 64, 128, 32, 2, 8>
   max err: 16
Perf:   0.207681 ms, 5.170149 TFlops, 41.02291 GB/s, DeviceGemm_Xdl_CShuffle<256, 64, 128, 32, 8, 8>
   max err: 16
test int8 TT
test int8 TN
test int8 NT
test int8 NN

</p>
</details>