@rosenrodt 2022-04-27
Post merge actions:
- `s_nop` padding on MI100
- CI test after mainline support for `setprio`, `sched_barrier()`: adding interwave pipeline kernels to device kernel pool (ckProfiler)
- CI test before mainline support for `setprio`, `sched_barrier()`: TBD